Add override method "listExtendConfig" to the ListController behavior
I have come up with this because in a custom ControllerBehavior that I am working I need to access the list configuration and make some changes like disabling sorting, disable showSetup, make a default sorting. I did not find a way to do this with the ListController), then I have taken as a reference the override function "relationExtendConfig" of the RelationController and made it for the ListController. So now I can do this in my custom behavior:
```
    public function listExtendConfig ($config, $definition = null)
    {
        $config->showSetup = false;
        $config->showSorting = false;
    }
```
I do not know if the function call (line 122) is the best place for it, but that worked like a charm for me (if there is a way to do it without this I would really like to know it :) ). Thanks in advance and sorry for my bad English.